requirements.txt
aws-lambda-powertools
awslambdaric
boto3
botocore
load-dotenv
pydantic